plugin-light-signal-detection: first skeleton for ground truth detection
Because vision in the simulation needs some effort and we might also
want to test in the simulation without sensor/perception errors, this
plugins should send the ground truth light signal detection to the robot
control software.
This is just a skeleton because we have no new refbox(-connection) and
therefore no light msgs.