mps-placement: randomly place static mps that do not shake under pucks
MPS were non-static before to be able to move them at the random
positions. This somehow made them shake when a product was placed on
them and caused products to fall down. Now MPS are staticly spawned
instead of moved to the new positions. This also speedsup the simulation
a bit.
I was not able to properly remove spawned MPS. (The plugin crashed or
only the visuals were removed.) Therefore, just reset the refbox (not
restart) when you want to start a new game.