plugins/puck: fix deadlock on model creation
The puck model announces itself with a NewPuck message. This triggers
action in the MPS model plugin which in turn tries to access the model
of the very puck just created. This access would be blocked by a
"loadModelMutex" in the Gazebo World, since the loading of the model
holds this at this time. At the same time, a transport-related lock
seems to be held. Therefore a classic deadlock stops the simulation.

This fixes the "laser data tool old" messages with a stopped simulation
clock and no help other than restarting. Maybe there are more such
causes, this is the most pressing and make things work for me again.