plugins/mps: remove wait time for tag positioning
Simply check if the tags have been spawned yet, and if so directly grab
the tags for the machines.

It would still be nicer to have tags to be attached to the MPS or even
an intrinsic part as to be able to move MPS with their tags in
simulation. Maybe another day... or student? ;-)