plugins/gripper: more general method to get gripper link
Because some teams included the gripper directly in the model and others
have the gripper in a submodel included in another model, I created a
more general method to get the gripper link. It just iterates over all
links of the model and searches for the gripper in the link name.