plugins: implement Gazebo 8 API changes
Lots of functions & classes are deprecated in Gazebo 8. This commit adds
a bunch of wrapper macros that redirect to the appropriate calls & types
depending on whether Gazebo V <= 5 or 5 < V < 8 or V >= 8. Yes, I know
this is an outrageous preprocessor hack, but trust me, you do not want
to see the corresponding amount of #if's and #else's in the source
files.