delivery station: wait for the puck and gate info before delivery
Previously, we always prepared the delivery statin and then put the work
piece into the machine. Now, the order can be changed, i.e., we can
first put the work piece into the machine and the prepare it. Do not
enforce any ordering in the simulation; instead, send the delivery
information as soon as the work piece is in the machine and the gate
info has been received.