delivery station: reset machine state on refbox msg
If the refbox sends a new machine state IDLE, also set the internal
state to IDLE. Otherwise, the DS always stays AVAILABLE after the first
delivery, which triggers the refbox to accept the delivery without a
puck (instead of waiting for the actual puck).

Also set the state of the DS to AVAILABLE at the right point, i.e., when
a puck was fed into the machine.