Refine failures at input and output of CapStation.
The input of CapStation can fail, by not accepting the dropped Pucks.
The output of CapStation can fail, by locking the puck it worked on.
The probabilities can be controlled in the config file. The break state
is redefined every rebreak-interval or when the machine receives the
BROKEN state. The interval is also controllable in the config file.

When "unbreaking" the input, all pucks there are removed. Same applies
to the output. Pucks in the shelf are not removed.