ring station: set AVAILABLE only if the MPS was prepared before
Before this change, AVAILABLE was set as soon as there was a work piece
in the MPS. This caused a race condition if the machine was not prepared
yet, i.e., if a robot first tries to put the work piece into the
machine, and only then prepare it. In this case, the refbox sends the
PREPARED state and immediately afterwards the PROCESSING state to the
MPS. In some cases, the information on the prepare is lost and the ring
station does not know which ring color to mount.

Now, we wait until we receive the prepare message before setting the
ring station to AVAILABLE. This way, we always get the prepare
information.