mps: process OPC-UA dispense base command
We need to dispense the base asynchronously, as otherwise the OPC-UA
server would no longer be responsive. Therefore, introduce a thread
object that always contains the command that is currently being
processed. Also make sure that we do not run two commands
simultaneously by ignoring the second command. This is slightly
different to the behavior on the real MPS, where the first command would
be skipped. Either way, the refbox should not send multiple commands to
the same MPS, therefore this should be fine.