mps: allow multiple workpieces in a machine
We no longer have a single `puck_in_processing_name_`, but instead three
workpieces `wp_in_input_`, `wp_in_middle_`, and `wp_in_output_`. This
allows us to have a workpiece in the input and at the same time another
workpiece in the output, as it often happens with cap stations and ring
stations.