mps: set busy to false only at the very end of conveyor move
If we set busy to false before resetting the action id and payload, we
may lose the next action, because the client already updated the action
id directly after we set busy to false. Instead, just move it to the
very end, so we first set ready to true (if appropriate) and then set
busy to false.

Also get rid of an unnecessary `sleep`.