delivery station: process command async, not in on_puck_msg handler
Processing the command directly in the handler freezes the simulation
for a few seconds, resulting in delocalized robots (as they do not get
any updates from the simulation for that time). Those commands should be
processed asynchronously, as we already do for all other commands. For
some reason, we did not do this for the delivery station.